+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 This changelog references the relevant changes (evolution, bug and security fixes)
 
+* 1.0.2
+    * Replace ComposerScriptHandler to InitializeInstallCommand
+    * Ask for console file path on command (prepare for Sf3)
+    * Add a note about removing useless IP check app_dev.php
+
 * 1.0.1
     
     * Add Grunt initialization command